RAG_Streamlit_App/requirements.txt
streamlit>=1.30.0
openai>=1.0.0
PyMuPDF>=1.23.0
spacy>=3.5.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
tenacity>=8.1.0
python-dotenv>=1.0.0
google-auth-oauthlib
google-auth
requests
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl